+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [NEXT_RELEASE]
+### Changed
+- [server] using multistage build on teresa dockerfile
+
 ## [0.14.0] - 2018-02-06
 ### Changed
 - [server] decouple k8s client interface from the domain ones