+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [server] using multistage build on teresa dockerfile
 - [server] update default slugbuilder version to v3.2.0
 
+### Added
+- `exec` command
+
 ## [0.14.0] - 2018-02-06
 ### Changed
 - [server] decouple k8s client interface from the domain ones