+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [server] upgrade golang version of server to 1.9
 - [helm] Add HealthChecks
 - [server] graceful shutdown
+- [client] support to deploy remote (http and https) and local files
 
 ## [0.14.0] - 2018-02-06
 ### Changed